added props type to blog

+++ Part 5/bloglist-frontend/src/components/Blog.js
@@ -1,28 +1,46 @@
 import Togglable from "./Togglable";
 import blogServices from "../services/blogs";
-const Blog = ({ setLiked,like, blog, setBlogs, blogs }) => {
+import PropTypes from 'prop-types'
+
+const Blog = ({ setLiked, like, blog, setBlogs, blogs }) => {
   const likeBlog = async () => {
     try {
       await blogServices.likeBlog(blog.id);
-      setLiked(!like)
+      setLiked(!like);
     } catch (e) {
       console.log(e);
     }
   };
   const deleteBlog = async () => {
-    try {
-      await blogServices.deleteBlog(blog.id);
-      setBlogs(() => blogs.filter((blogR) => blogR.id !== blog.id))
-    } catch (e) {
-      console.log(e);
+    if (window.confirm("Are you sure you want to delete this blog? ")) {
+      try {
+        await blogServices.deleteBlog(blog.id);
+        setBlogs(() => blogs.filter((blogR) => blogR.id !== blog.id));
+      } catch (e) {
+        if (e.response?.data?.error) {
+          alert(e.response.data.error);
+        }
+      }
     }
   };
   return (
     <div style={{ flexDirection: "column", cursor: "default" }}>
-      <span style={{ display: "flex", alignItems:"center", justifyItems:"center" }}>
+      <span
+        style={{
+          display: "flex",
+          alignItems: "center",
+          justifyItems: "center",
+        }}
+      >
         <p>{blog.title}</p>
         <button
-          style={{ color: "blue", textAlign: "center", borderRadius: "8px", height:"min-content", marginLeft:"10px" }}
+          style={{
+            color: "blue",
+            textAlign: "center",
+            borderRadius: "8px",
+            height: "min-content",
+            marginLeft: "10px",
+          }}
           onClick={deleteBlog}
         >
           Delete
@@ -31,10 +49,22 @@ const Blog = ({ setLiked,like, blog, setBlogs, blogs }) => {
 
       <Togglable buttonLabel="View">
         <p>{"URL: " + blog.url}</p>
-        <span style={{ display: "flex", alignItems:"center", justifyItems:"center"  }}>
+        <span
+          style={{
+            display: "flex",
+            alignItems: "center",
+            justifyItems: "center",
+          }}
+        >
           <p>{"Likes: " + blog.likes}</p>
           <button
-            style={{ color: "blue", textAlign: "center", borderRadius: "8px", height:"min-content", marginLeft:"10px", }}
+            style={{
+              color: "blue",
+              textAlign: "center",
+              borderRadius: "8px",
+              height: "min-content",
+              marginLeft: "10px",
+            }}
             onClick={likeBlog}
           >
             Like
@@ -46,5 +76,13 @@ const Blog = ({ setLiked,like, blog, setBlogs, blogs }) => {
     </div>
   );
 };
+
+Blog.propTypes = {
+  blog : PropTypes.object.isRequired,
+  setBlogs : PropTypes.func.isRequired,
+  blogs : PropTypes.array.isRequired,
+  like : PropTypes.bool.isRequired,
+  setLiked : PropTypes.func.isRequired
+}
 
 export default Blog;
